+++ MMM/Sources/Reactor/PushSettingReactor.swift
@@ -82,9 +82,7 @@ extension PushSettingReactor {
             return Observable.concat([
                 .just(.setPresentSheetView(true)),
                 .just(.setPresentSheetView(false))])
-            //        case .newsPushSwitchToggle(let isOn):
-            //            let request = PushAgreeUpdateReqDto(pushAgreeDvcd: "01", pushAgreeYN: isOn ? "Y" : "N")
-            //            return pushAgreeUpdate(request)
+
         case .newsPushSwitchToggle(let isOn):
             let request = PushAgreeUpdateReqDto(pushAgreeDvcd: "01", pushAgreeYN: isOn ? "Y" : "N")
             return Observable.concat([
@@ -95,22 +93,7 @@ extension PushSettingReactor {
             let notificationCenter = UNUserNotificationCenter.current()
             
             if isOn {
-                notificationCenter.requestAuthorization(options: [.alert, .sound, .badge]) { granted, error in
-                    if granted {
-                        print("알림 허용")
-                        
-                        let time = Common.getCustomPushTime()
-                        let weekList = Common.getCustomPushWeekList()
-                        for (day, isOn) in weekList.enumerated() {
-                            if isOn {
-                                // 특정 요일 선택 (1은 일요일, 2는 월요일, ..., 7은 토요일)
-                                self.scheduleWeeklyNotification(day: day+1, time: time)
-                            }
-                        }
-                    } else {
-                        print("알림 거부")
-                    }
-                }
+
             } else {
                 notificationCenter.removeAllPendingNotificationRequests()
             }
@@ -151,6 +134,9 @@ extension PushSettingReactor {
             }
             
         case .updatePushAgreeSwitch(let response, let error):
+            if let error = error {
+                print(error.localizedDescription)
+            }
             print(response)
 
         // MARK: - 여긴 완성
@@ -183,47 +169,5 @@ extension PushSettingReactor {
             .map { (response, error) -> Mutation in
                 return .updatePushAgreeSwitch(response, error)
             }
-    }
-        
-    func scheduleWeeklyNotification(day: Int, time: String) {
-        // 알림 내용 설정
-        let notificationCenter = UNUserNotificationCenter.current()
-        
-//        notificationCenter.removeAllPendingNotificationRequests()
-        
-        let hour = time.components(separatedBy: ":")[0]
-        let minute = time
-            .components(separatedBy: ":")[1]
-            .components(separatedBy: " ")[0]
-        
-        let content = UNMutableNotificationContent()
-        content.title = "MMM"
-        content.body = Common.getCustomPushText()
-        
-        // 알림 발생 시간 설정
-        var dateComponents = DateComponents()
-        
-        dateComponents.hour = Int(hour)! // 알림을 원하는 시간으로 변경
-        dateComponents.minute = Int(minute)!
-
-        dateComponents.weekday = day // 특정 요일 선택 (1은 일요일, 2는 월요일, ..., 7은 토요일)
-        // 알림 요청 생성
-        let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: true)
-        
-        let request = UNNotificationRequest(identifier: "weeklyNoti\(day)", content: content, trigger: trigger)
-
-        // 알림 요청 등록
-        notificationCenter.add(request) { (error) in
-            if let error = error {
-                print("알림 등록 중 오류 발생: \(error.localizedDescription)")
-            } else {
-                print("알림이 성공적으로 등록되었습니다.")
-                UNUserNotificationCenter.current().getPendingNotificationRequests { requests in
-                    for request in requests {
-                        print("예약된 알림: \(request.identifier)")
-                    }
-                }
-            }
-        }
     }
 }
